backend/requirements.txt
flask
qdrant-client
openai
voyageai
pdfplumber
requests
bs4
flask_cors
python-dotenv